feat(runner): create `ConnectionManager`
